Added queue persistence and priority ordering

+++ Migrations/201809230209393_Init.cs
@@ -12,7 +12,7 @@ namespace NFive.Queue.Migrations
                 c => new
                     {
                         Id = c.Guid(nullable: false),
-                        Priority = c.Short(nullable: false),
+                        Priority = c.Int(nullable: false),
                         Position = c.Short(nullable: false),
                         SessionId = c.Guid(nullable: false),
                         Created = c.DateTime(nullable: false, precision: 0),
@@ -26,7 +26,7 @@ namespace NFive.Queue.Migrations
         public override void Down()
         {
             DropForeignKey("dbo.QueuePlayers", "SessionId", "dbo.Sessions");
-            DropIndex("dbo.QueuePlayers", new[] { "SessionId" });
+	        DropIndex("dbo.QueuePlayers", new[] { "SessionId" });
             DropTable("dbo.QueuePlayers");
         }
     }
